Gcc warnings ptr to int

+++ src/core/templates.h
@@ -127,9 +127,9 @@ public:
 		int i = GetJustIndex(entry);
 		return m_flags[i].u + (i<<8);
 	}
-	int GetJustIndex(T *entry){
+	uintptr GetJustIndex(T *entry){
 		// TODO: the cast is unsafe
-		return (int)((U*)entry - m_entries);
+		return (uintptr)((U*)entry - m_entries);
 	}
 	int GetNoOfUsedSpaces(void) const{
 		int i;
